Make sure app load does not fail for lack of Domain table
Rake needs to successfully load the app before it will migrate the
database.  Whether or not this is the correct behavior is up for
debate, but in the mean time we will leave off serving multidomains
in the rack config file if there is no database table. This solves
the chicken-and-egg problem for `rake db:automigrate`.